requirements.txt
asyncio==3.4.3
python-dotenv==1.0.1
supabase==2.5.1
supabase_client==0.2.5